Apimenov quad

+++ core/joystick/playbackJoystickInterface.cpp
@@ -1,5 +1,6 @@
 #include <fstream>
 #include <unistd.h>
+#include <sstream>
 
 #include "core/utils/global.h"
 #include "core/utils/utils.h"
@@ -16,7 +17,7 @@ PlaybackJoystickInterface::PlaybackJoystickInterface(const string &deviceName):
     corecvs::JoystickInterface(deviceName)
 {
     data.load(deviceName);
-    SYNC_PRINT(("PlaybackJoystickInterface::PlaybackJoystickInterface(): Loaded %d joystick moves\n", data.states.size()));
+    SYNC_PRINT(("PlaybackJoystickInterface::PlaybackJoystickInterface(): Loaded %d joystick moves\n", (int)data.states.size()));
 }
 
 std::vector<string> PlaybackJoystickInterface::getDevices(const string &prefix)
